Change starter code return types to usize
A reasonable default because it slightly reduces friction with Rust
APIs.

+++ src/days/day01.rs
@@ -1,6 +1,6 @@
 use crate::harness::input::RawInput;
 
-pub fn solve_part1(input: RawInput) -> u64 {
+pub fn solve_part1(input: RawInput) -> usize {
     let (mut lefts, mut rights) = get_lists(input);
     lefts.sort();
     rights.sort();
@@ -11,19 +11,19 @@ pub fn solve_part1(input: RawInput) -> u64 {
         .sum()
 }
 
-pub fn solve_part2(input: RawInput) -> u64 {
+pub fn solve_part2(input: RawInput) -> usize {
     let (lefts, rights) = get_lists(input);
     lefts
         .into_iter()
-        .map(|left| rights.iter().filter(|&&right| right == left).count() as u64 * left)
+        .map(|left| rights.iter().filter(|&&right| right == left).count() * left)
         .sum()
 }
 
-fn get_lists(input: RawInput) -> (Vec<u64>, Vec<u64>) {
+fn get_lists(input: RawInput) -> (Vec<usize>, Vec<usize>) {
     let mut lefts = vec![];
     let mut rights = vec![];
     input
-        .per_line(|line| line.split_whitespace::<u64>())
+        .per_line(|line| line.split_whitespace::<usize>())
         .for_each(|nums| {
             lefts.push(nums[0]);
             rights.push(nums[1]);
