+++ src/days/day23.rs
@@ -1,9 +1,9 @@
 use crate::harness::input::RawInput;
 
-pub fn solve_part1(input: RawInput) -> u64 {
+pub fn solve_part1(input: RawInput) -> usize {
     todo!("{}", &input.as_str()[..0])
 }
 
-pub fn solve_part2(input: RawInput) -> u64 {
+pub fn solve_part2(input: RawInput) -> usize {
     todo!("{}", &input.as_str()[..0])
 }
